Split API routes by domain
