+++ tests/test_views/test_api/test_jobs.py
@@ -1,0 +1,3 @@
+class TestGetJobStatus:
+    def test_something(self, client):
+        pass
